requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.4
flask-wtf==0.14.3
pip-tools==5.3.1
black==20.8b1
connexion[swagger-ui]